N2G/requirements.txt
crewai>=0.20.0
crewai-tools>=0.1.3
streamlit>=1.32.0
openai>=1.0.0
python-dotenv
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
